+++ src/lib/native/NativeController.js
@@ -4,6 +4,7 @@ import Cryptography from '../Crypto'
 import NativeAccountStorage from './NativeAccountStorage'
 import Account from '../Account'
 import { STATUS_ALLGOOD, STATUS_DISABLED, STATUS_ERROR, STATUS_SYNCING } from '../interfaces/Controller'
+import { freeStorageIfNecessary } from '../IndexedDB'
 
 const INACTIVITY_TIMEOUT = 1000 * 7
 const MAX_BACKOFF_INTERVAL = 1000 * 60 * 60 // 1 hour
@@ -35,12 +36,12 @@ class AlarmManager {
       const lastSync = data.lastSync || 0
       const interval = data.syncInterval || DEFAULT_SYNC_INTERVAL
       if (data.scheduled) {
-        this.ctl.scheduleSync(accountId)
+        await this.ctl.scheduleSync(accountId)
         continue
       }
       if (data.error && data.errorCount > 1) {
         if (Date.now() > this.getBackoffInterval(interval, data.errorCount, lastSync) + lastSync) {
-          this.ctl.scheduleSync(accountId)
+          await this.ctl.scheduleSync(accountId)
           continue
         }
         continue
@@ -49,7 +50,7 @@ class AlarmManager {
         Date.now() >
         interval * 1000 * 60 + data.lastSync
       ) {
-        this.ctl.scheduleSync(accountId)
+        await this.ctl.scheduleSync(accountId)
       }
     }
   }
@@ -82,6 +83,17 @@ export default class NativeController {
     this.listeners = []
 
     this.alarms = new AlarmManager(this)
+
+    // Remove old logs
+
+    NativeAccountStorage.changeEntry(
+      'logs',
+      log => {
+        return []
+      },
+      []
+    )
+    freeStorageIfNecessary()
 
     // lock accounts when locking is enabled
 
@@ -133,6 +145,7 @@ export default class NativeController {
       if (this.schedule[accountId]) {
         clearTimeout(this.schedule[accountId])
       }
+      console.log('scheduleSync: setting a timeout in ms :', INACTIVITY_TIMEOUT)
       this.schedule[accountId] = setTimeout(
         () => this.scheduleSync(accountId),
         INACTIVITY_TIMEOUT